Add list parser module
The list_parser.py here is mostly a line by line translation of same
functions in MW's parser.php. Amazingly, it worked fine for some cases,
but unfortunately not for others. There are perhaps some bugs and besides
that the HTML output needs to be parsed again.

This module may be completely rewritten or removed in the future.
This commit is solely to keep a copy of the work.