version='0.9.0'
I wanted to add tag searching methods in this version but lots of new
features and bug fixes were added in between and I don't know when I
will be able to finish tag related functions.
There were lots of changes that I wanted to release, and here they are!
Enjoy!