docs(README.rst): note that there is no walk method
This is mostly due to the lasy nature of wikitextparser. Many nodes
are not concrete, not until the first call to their respective fetchers.
For example there is no external link node in the initial tree (which
is actually not a tree yet (for performance reasons which I'm not
sure still hold or not)). However, lately I've been thinking more
and more about implementing such a tree.

Closes #35