fix(wikilink): empty wikilinks are internally not replaced with _
"[[|alt]]" is a valid wikilink, but WIKILINK_PARAM_FINDITER did not
consider it valid. In result, it was seens as a wikilink, but not
marked as _ for, as example, a parser function to handle it. This
meant a parser function saw "alt ]]" as another argument, where it
really was not.